Refactor API structure and add middleware, health check
Refactored the API to modularize configuration, middleware, and health check endpoint. Moved application and config structs to a new config.go file, added middleware.go for authentication and rate limiting, and introduced a health check handler. Updated main.go to use the new structure, improved environment variable usage, and enhanced error handling. Fixed type inconsistencies in models and seeding logic. Cleaned up unused code and improved maintainability.

+++ internal/store/posts.go
@@ -8,7 +8,7 @@ import (
 )
 
 type Post struct {
-	ID        uint           `gorm:"primaryKey" json:"id"`
+	ID        int64           `gorm:"primaryKey" json:"id"`
 	Title     string         `gorm:"size:255" json:"title"`
 	Content   string         `json:"content"`
 	UserID    uint           `json:"user_id"`
